test fixture location
