+++ test/__fixtures__/holder_0_private_key.cbor
