+++ test/__fixtures__/holder_0_public_key.cbor
