+++ test/__fixtures__/issuer_0_private_key.json
