+++ test/__fixtures__/issuer_0_public_key.json
