+++ test/__fixtures__/minimal_credential.yml
