+++ test/__fixtures__/minimal_credential_with_bad_urls.yml
