+++ test/__fixtures__/minimal_credential_with_dids.yml
